src/requirements.txt
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-python-client==2.166.0
python-dotenv==1.1.0
youtube-transcript-api==1.0.3
openai==1.72.0
pydantic==2.11.3
fsrs==5.1.2
dj-database-url
psycopg2-binary
gunicorn
whitenoise
python-decouple
stripe==12.0.0
django-guest-user==0.5.5
lucide==1.1.3
